wip: proxyd cl layer 2
